final_pjt/Movie-Community/Django-back/requirements.txt
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.3.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
joblib==0.15.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
psycopg2==2.8.5
pylint==2.5.3
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
